requirements.txt
FastAPI
uvicorn
automata-lib
pydantic
```